Add delete modal for warehouse

+++ saleor/dashboard/warehouse/views.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING
 
 from django.conf import settings
 from django.contrib import messages
@@ -7,10 +7,13 @@ from django.shortcuts import get_object_or_404, redirect
 from django.template.response import TemplateResponse
 from django.utils.translation import pgettext_lazy
 
-from saleor.account.models import Address
 from saleor.core.utils import get_paginator_items
 from saleor.dashboard.views import staff_member_required
-from saleor.dashboard.warehouse.forms import WarehouseAddressForm, WarehouseForm
+from saleor.dashboard.warehouse.forms import (
+    WarehouseAddressForm,
+    WarehouseForm,
+    save_warehouse_from_forms,
+)
 from saleor.warehouse.models import Warehouse
 
 if TYPE_CHECKING:
@@ -31,29 +34,30 @@ def index(request: "HttpRequest") -> "HttpResponse":
 
 @staff_member_required
 @permission_required("warehouse.manage_warehouses")
-def warehouse_form(
-    request: "HttpRequest", uuid: Optional["UUID"] = None
-) -> "HttpResponse":
-    if uuid is not None:
-        qs = Warehouse.objects.select_related("address").prefetch_related(
-            "shipping_zones"
-        )
-        warehouse = get_object_or_404(qs, pk=uuid)
-        address = warehouse.address
-    else:
-        warehouse = Warehouse()
-        address = Address()
-
-    warehouse_form = WarehouseForm(request.POST or None, instance=warehouse)
-    address_form = WarehouseAddressForm(request.POST or None, instance=address)
+def warehouse_create(request: "HttpRequest") -> "HttpResponse":
+    warehouse_form = WarehouseForm(request.POST or None)
+    address_form = WarehouseAddressForm(request.POST or None)
     if address_form.is_valid() and warehouse_form.is_valid():
-        address = address_form.save()
-        warehouse_form.save_with_address(address)
+        warehouse = save_warehouse_from_forms(warehouse_form, address_form)
+        msg = pgettext_lazy("Dashboard message", "Warehouse created")
+        messages.success(request, msg)
+        return redirect("dashboard:warehouse-detail", uuid=warehouse.uuid)
+    ctx = {"warehouse_form": warehouse_form, "address_form": address_form}
+    return TemplateResponse(request, "dashboard/warehouse/form.html", ctx)
 
-        if uuid is not None:
-            msg = pgettext_lazy("Dashboard message", "Warehouse updated")
-        else:
-            msg = pgettext_lazy("Dashboard message", "Warehouse created")
+
+@staff_member_required
+@permission_required("warehouse.manage_warehouses")
+def warehouse_update(request: "HttpRequest", uuid: "UUID") -> "HttpResponse":
+    qs = Warehouse.objects.select_related("address").prefetch_related("shipping_zones")
+    warehouse = get_object_or_404(qs, pk=uuid)
+    warehouse_form = WarehouseForm(request.POST or None, instance=warehouse)
+    address_form = WarehouseAddressForm(
+        request.POST or None, instance=warehouse.address
+    )
+    if address_form.is_valid() and warehouse_form.is_valid():
+        save_warehouse_from_forms(warehouse_form, address_form)
+        msg = pgettext_lazy("Dashboard message", "Warehouse updated")
         messages.success(request, msg)
         return redirect("dashboard:warehouse-detail", uuid=warehouse.uuid)
     ctx = {
@@ -76,7 +80,12 @@ def warehouse(request: "HttpRequest", uuid: "UUID") -> "HttpResponse":
 @permission_required("warehouse.manage_warehouses")
 def warehouse_delete(request: "HttpRequest", uuid: "UUID") -> "HttpResponse":
     warehouse = get_object_or_404(Warehouse, pk=uuid)
-    warehouse.delete()
-    msg = pgettext_lazy("Dashboard message", "Warehouse deleted")
-    messages.success(request, msg)
-    return redirect("dashboard:warehouse-list")
+    if request.method == "POST":
+        warehouse.delete()
+        msg = pgettext_lazy("Dashboard message", "Warehouse deleted")
+        messages.success(request, msg)
+        return redirect("dashboard:warehouse-list")
+    ctx = {"warehouse": warehouse}
+    return TemplateResponse(
+        request, "dashboard/warehouse/modal/confirm_delete.html", ctx
+    )
